PoC: Generate SPIFFE identities in csi-driver
This is inteded as a quick demonstration of a potential replacement for
csi-driver-spiffe. This isn't official, isn't on the roadmap and there's
no guarantee that anything here will ever be used in any format.

This is intended as a point of discussion following recent discussions
between maintainers and in the cert-manager daily standup

Signed-off-by: Ashley Davis <ashley.davis@venafi.com>